requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic_settings
langchain
groq
langchain-groq
mangum